src/requirements/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
Pillow==8.4.0
celery==5.2.7
channels==4.0.0
asgiref==3.5.0
python-dateutil==2.8.2
PyJWT==2.4.0
django_celery_beat
django_cleanup
djoser
django-taggit
channels-redis
celery[redis]
python-dotenv~=1.0.0
django_rest_passwordreset
pymysql
django-axes
django-axes[ipware]
g4f
django-cors-headers
python-slugify == 8.0.1
moviepy
pillow
ushlex
bing_image_downloader
uuid
pyttsx3
pytubefix
TTS
face_alignment==1.3.5
imageio==2.19.3
imageio-ffmpeg==0.4.7
librosa==0.9.2 #
numba
resampy==0.3.1
pydub==0.25.1
scipy==1.10.1
kornia==0.6.8
tqdm
yacs==0.1.8
pyyaml
joblib==1.1.0
scikit-image==0.19.3
basicsr==1.4.2
facexlib==0.3.0
gradio
gfpgan
av
safetensors
drf-yasg
llvmlite==0.38.1
openai
google-generativeai
anthropic
django-resized